etl-pipeline/pyproject.toml
[tool.poetry]
name = "etl-pipeline"
version = "0.1.0"
description = ""
authors = ["Hau Ho Dang <harima1301@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
psycopg2-binary = "2.9.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
